+++ appointmentManagementAPI/src/main/java/com/Hbence/appointmentManagementAPI/repository/DeviceCategoryRepository.java
@@ -3,5 +3,5 @@ package com.Hbence.appointmentManagementAPI.repository;
 import com.Hbence.appointmentManagementAPI.entity.DevicesCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeviceCategoryRepository extends JpaRepository<DevicesCategory, Integer> {
+public interface DeviceCategoryRepository extends JpaRepository<DevicesCategory, Long> {
 }
