+++ appointmentManagementAPI/src/main/java/com/Hbence/appointmentManagementAPI/repository/DeviceRepository.java
@@ -3,5 +3,5 @@ package com.Hbence.appointmentManagementAPI.repository;
 import com.Hbence.appointmentManagementAPI.entity.Devices;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeviceRepository extends JpaRepository<Devices, Integer> {
+public interface DeviceRepository extends JpaRepository<Devices, Long> {
 }
