+++ appointmentManagementAPI/src/main/java/com/Hbence/appointmentManagementAPI/repository/GalleryRepository.java
@@ -3,5 +3,5 @@ package com.Hbence.appointmentManagementAPI.repository;
 import com.Hbence.appointmentManagementAPI.entity.Gallery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GalleryRepository extends JpaRepository<Gallery, Integer> {
+public interface GalleryRepository extends JpaRepository<Gallery, Long> {
 }
