+++ appointmentManagementAPI/src/main/java/com/Hbence/appointmentManagementAPI/repository/HistoryRepository.java
@@ -3,5 +3,5 @@ package com.Hbence.appointmentManagementAPI.repository;
 import com.Hbence.appointmentManagementAPI.entity.History;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HistoryRepository extends JpaRepository<History, Integer> {
+public interface HistoryRepository extends JpaRepository<History, Long> {
 }
