+++ appointmentManagementAPI/src/main/java/com/Hbence/appointmentManagementAPI/repository/NewsRepository.java
@@ -2,8 +2,6 @@ package com.Hbence.appointmentManagementAPI.repository;
 
 import com.Hbence.appointmentManagementAPI.entity.News;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.query.Procedure;
-import org.springframework.data.repository.query.Param;
 
-public interface NewsRepository extends JpaRepository <News, Integer> {
+public interface NewsRepository extends JpaRepository <News, Long> {
 }
