+++ appointmentManagementAPI/src/main/java/com/Hbence/appointmentManagementAPI/repository/PaymentMethodRepository.java
@@ -3,5 +3,5 @@ package com.Hbence.appointmentManagementAPI.repository;
 import com.Hbence.appointmentManagementAPI.entity.PaymentMethods;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentMethodRepository extends JpaRepository<PaymentMethods, Integer> {
+public interface PaymentMethodRepository extends JpaRepository<PaymentMethods, Long> {
 }
