+++ appointmentManagementAPI/src/main/java/com/Hbence/appointmentManagementAPI/repository/RuleRepository.java
@@ -3,5 +3,5 @@ package com.Hbence.appointmentManagementAPI.repository;
 import com.Hbence.appointmentManagementAPI.entity.Rules;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RuleRepository extends JpaRepository<Rules, Integer> {
+public interface RuleRepository extends JpaRepository<Rules, Long> {
 }
