+++ appointmentManagementAPI/src/main/java/com/Hbence/appointmentManagementAPI/repository/SpecialOfferRepository.java
@@ -3,5 +3,5 @@ package com.Hbence.appointmentManagementAPI.repository;
 import com.Hbence.appointmentManagementAPI.entity.SpecialOffer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpecialOfferRepository extends JpaRepository <SpecialOffer, Integer> {
+public interface SpecialOfferRepository extends JpaRepository <SpecialOffer, Long> {
 }
